pyproject.toml
[project]
name = "kindle"
version = "0.1.0"
description = "kindle"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "alembic>=1.16.4",
    "black>=25.1.0",
    "pre-commit>=4.3.0",
    "ruff>=0.12.8",
]

test = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
]

[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.black]
line-length = 90
target-version = ["py311"]
